Added sp_GetDDL.sql
The sp is taken from https://www.stormrage.com/SQLStuff/sp_GetDDL_Latest.txt

Unfortunately no contact was found at https://www.stormrage.com in order to refer to the true source of this sp. So, I added the sp to this repository so that developers can work on and improve it.